Functional save locations by userID, Fixed tracking proper id through login/logout

+++ Roamio/screens/AccountScreen.js
@@ -1,75 +1,39 @@
-import React, {useEffect, useState} from 'react';
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+import React, { useEffect, useState } from 'react';
+import { View, Text, StyleSheet, TouchableOpacity, Alert } from 'react-native';
 import axios from "axios";
-import {useIsFocused, useNavigation, useRoute} from "@react-navigation/native";
+import { useIsFocused, useNavigation } from "@react-navigation/native";
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-const AccountScreen = ({ navigation }) => {
-    //const navigation = useNavigation();
+const AccountScreen = () => {
+    const navigation = useNavigation();
+    const isFocused = useIsFocused();
+
     const [loggedIn, setLoggedIn] = useState(false);
-   // const { userID } = route.params || {};
-    //const [userIDState, setUserIDState] = useState(null);
     const [userID, setUserID] = useState(null);
 
     useEffect(() => {
-        // Check if userID exists in AsyncStorage
-        retrieveUserID();
-    }, []);
-
-    const storeUserID = async (userID) => {
-        try {
-            const userIDString = JSON.stringify(userID); // Stringify the userID
-            await AsyncStorage.setItem('userID', userIDString);
-            setUserID(userID);
-        } catch (error) {
-            console.error('Error storing userID:', error);
+        // Check login status when the screen is focused
+        if (isFocused) {
+            checkLoginStatus();
+            retrieveUserID();
         }
-    };
+    }, [isFocused]);
 
     const retrieveUserID = async () => {
         try {
             const storedUserID = await AsyncStorage.getItem('userID');
             if (storedUserID !== null) {
-                // Parse the stored userID back to its original format
-                const userID = JSON.parse(storedUserID);
-                // userID found in AsyncStorage
-                setUserID(userID);
+                setUserID(JSON.parse(storedUserID));
             }
         } catch (error) {
             console.error('Error retrieving userID:', error);
         }
     };
 
-
-
-    const handleRegisterPress = () => {
-        navigation.navigate('Register');
-    };
-
-    const handleLoginPress = () => {
-        navigation.navigate('Login');
-    };
-
-
-
-    const isFocused = useIsFocused(); //used to check if the current screen is the one loaded
-
-    //navigation.navigate('Main', { userID });
-    useEffect(() => {
-        if (isFocused) {
-            checkLoginStatus(); // Call checkLoginStatus only when the screen is focused
-            //navigation.navigate('Main', { userID });
-            //setUserIDState(userID);
-           // console.log(userID)
-        }
-   // }, [isFocused, userID]); // Add isFocused to the dependency array
-    }, [isFocused]); // Add isFocused to the dependency array
-
     const checkLoginStatus = async () => {
         try {
             const response = await axios.get('http://192.168.1.241:3000/checkLoginStatus');
             setLoggedIn(response.data.loggedIn);
-           // console.log("status",loggedIn)
         } catch (error) {
             console.error('Error checking login status:', error);
         }
@@ -77,29 +41,26 @@ const AccountScreen = ({ navigation }) => {
 
     const handleLogout = async () => {
         try {
-            //console.log("logging out")
             const response = await axios.get('http://192.168.1.241:3000/logout');
-            setLoggedIn(response.data.loggedIn);
-            await AsyncStorage.removeItem('userID');
-            await AsyncStorage.clear();
-            // Get all keys from AsyncStorage
-            const getAllKeys = async () => {
-                let keys = [];
-                try {
-                    keys = await AsyncStorage.getAllKeys();
-                    console.log(keys);
-                } catch (e) {
-                    console.error('Error reading AsyncStorage keys:', e);
-                }
-            };
-
-            // Call getAllKeys function
-            await getAllKeys();
-            //console.log(userID)
-            //console.log("status", loggedIn)
+            if (response.data.loggedOut) {
+                await AsyncStorage.clear();
+                setLoggedIn(false);
+                setUserID(null);
+            } else {
+                Alert.alert('Logout Failed', 'Failed to logout. Please try again.');
+            }
         } catch (error) {
             console.error('Error logging out:', error);
+            Alert.alert('Error', 'An error occurred while logging out. Please try again.');
         }
+    };
+
+    const handleRegisterPress = () => {
+        navigation.navigate('Register');
+    };
+
+    const handleLoginPress = () => {
+        navigation.navigate('Login');
     };
 
     return (
